+++ src/main/java/com/example/bankcards/entity/Card.java
@@ -6,6 +6,7 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "cards", schema = "bankrest")
@@ -39,7 +40,7 @@ public class Card {
 
     @NotNull
     @Column(name = "balance", nullable = false, precision = 19, scale = 2)
-    private double balance;
+    private BigDecimal balance;
 
     @NotNull
     @Size(min = 3, max = 3, message = "CVV должен содержать 3 цифры")
@@ -56,7 +57,7 @@ public class Card {
     public Card() {
     }
 
-    public Card(String number, String name, String expiration, CardStatus status, double balance, String cvv, User user) {
+    public Card(String number, String name, String expiration, CardStatus status, BigDecimal balance, String cvv, User user) {
         this.number = number; // изменил ИИ: удалил вызов EncryptionUtil.encrypt, так как шифрование теперь обрабатывается конвертером (соответствие SOLID: OCP - изменения без модификации сущности).
         this.name = name;
         this.expiration = expiration;
@@ -106,11 +107,11 @@ public class Card {
         this.status = status;
     }
 
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 
-    public void setBalance(double balance) {
+    public void setBalance(BigDecimal balance) {
         this.balance = balance;
     }
