+++ src/main/java/com/example/bankcards/service/TransactionService.java
@@ -2,8 +2,10 @@ package com.example.bankcards.service;
 
 import com.example.bankcards.entity.Transaction;
 
+import java.math.BigDecimal;
+
 // добавил: Интерфейс для сервиса переводов между картами
 public interface TransactionService {
     // добавил: Метод для выполнения перевода
-    Transaction transfer(Long fromCardId, Long toCardId, double amount, Long userId);
+    Transaction transfer(Long fromCardId, Long toCardId, BigDecimal amount, Long userId);
 }
